Add Queue Critical Press Exercise
Added exercise where a queue is used to send the change of state of two buttons from two tasks that perform the debounce. In this exercise we check the critical condition (button not pressed) and if it is active we force the sending to the queue in case it is full.